+++ Common/src/main/java/de/markusbordihn/dynamicprogressiondifficulty/data/ItemClassData.java
@@ -31,12 +31,16 @@ import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.item.ArmorStandItem;
 import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.BoatItem;
+import net.minecraft.world.item.DispensibleContainerItem;
 import net.minecraft.world.item.DyeItem;
+import net.minecraft.world.item.HangingEntityItem;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemFrameItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.item.MinecartItem;
 import net.minecraft.world.item.RecordItem;
 import net.minecraft.world.item.SpawnEggItem;
 import org.apache.logging.log4j.LogManager;
@@ -71,20 +75,11 @@ public class ItemClassData {
       ResourceKey<Item> resourceKey = itemEntry.getKey();
       Item item = itemEntry.getValue();
 
-      // Ignore items which are not enabled or are not relevant for the item class mapping.
-      if (item == null
-          || item.equals(Items.AIR)
-          || item instanceof BlockItem
-          || item instanceof SpawnEggItem
-          || item instanceof DyeItem
-          || item instanceof ItemFrameItem
-          || item instanceof RecordItem
-          || item.isEdible()
-          || item.getDefaultInstance().is(ModItemTags.IGNORED_ITEMS)) {
+      // Ignore items which are ignored or are not relevant for the item class mapping.
+      String itemName = resourceKey.location().toString();
+      if (!isRelevantItem(item, itemName)) {
         continue;
       }
-
-      String itemName = resourceKey.location().toString();
 
       // 1. Mapping items to categorize by tags.
       ItemClass tagItemClass = ItemClassHelper.getItemClassByTag(item);
@@ -97,6 +92,14 @@ public class ItemClassData {
       }
 
       // 2. Mapping items by name
+      ItemClass nameItemClass = ItemClassHelper.getItemClassByItemName(itemName);
+      if (nameItemClass != null) {
+        log.info("Item {} was mapped to category {} by name.", itemName, nameItemClass);
+        itemClassMap.put(item, nameItemClass);
+        itemClassItems.get(nameItemClass).add(item);
+        nameItemClass.setEnabled(true);
+        continue;
+      }
 
       // 3. Mapping items by item instance
       ItemClass instanceItemClass = ItemClassHelper.getItemClassByItemInstance(item);
@@ -115,6 +118,46 @@ public class ItemClassData {
     log.info(
         "Finished mapping items to categories in {} ms.", System.currentTimeMillis() - startTime);
     isRegistered = true;
+  }
+
+  public static boolean isRelevantItem(Item item, String itemName) {
+    return isRelevantItem(item) && isRelevantItem(itemName);
+  }
+
+  public static boolean isRelevantItem(Item item) {
+    return item != null
+        && !item.equals(Items.AIR)
+        && !(item instanceof BlockItem)
+        && !(item instanceof HangingEntityItem)
+        && !(item instanceof SpawnEggItem)
+        && !(item instanceof DyeItem)
+        && !(item instanceof RecordItem)
+        && !(item instanceof BoatItem)
+        && !(item instanceof MinecartItem)
+        && !(item instanceof ArmorStandItem)
+        && !(item instanceof DispensibleContainerItem)
+        && !item.isEdible()
+        && !item.getDefaultInstance().is(ModItemTags.IGNORED_ITEMS);
+  }
+
+  public static boolean isRelevantItem(String itemName) {
+    return itemName != null
+        && !itemName.isEmpty()
+        && !itemName.endsWith("_blueprint")
+        && !itemName.endsWith("_cast")
+        && !itemName.endsWith("_decoration")
+        && !itemName.endsWith("_grip")
+        && !itemName.endsWith("_handle")
+        && !itemName.endsWith("_head")
+        && !itemName.endsWith("_ingot")
+        && !itemName.endsWith("_limb")
+        && !itemName.endsWith("_nugget")
+        && !itemName.endsWith("_part")
+        && !itemName.endsWith("_pattern")
+        && !itemName.endsWith("_repair_kit")
+        && !itemName.endsWith("_template")
+        && !itemName.endsWith("_upgrade")
+        && !itemName.endsWith("_upgrade_kit");
   }
 
   public static ItemClass getItemClass(ItemStack itemStack) {